Data Management & UI update
Resolved several bugs in relation to saving inputted data from the user in cases of returning from a notification or from an activity. Also included a draft for a camera usage feature. Not fully operational yet. Also updated the User interface for better viewing in the DisplayList activity. Plan on working on the UI a bit more and testing the app for any bugs.